refactor: 북마크 버튼 컴포넌트화
